converte o resultado da divisão em decimal

+++ calculadora/src/br/com/dio/Calculadora.java
@@ -16,7 +16,7 @@ public class Calculadora {
         int somar = somar(a, b);
         int subtrair = subtrair(a, b);
         int multiplicar = multiplicar(a, b);
-        int dividir = dividir(a,b);
+        float dividir = dividir(a,b);
 
         System.out.println("somar: " + somar);
         System.out.println("subt: " + subtrair);
@@ -34,8 +34,8 @@ public class Calculadora {
     public static int multiplicar(int a, int b) {
         return a * b;
     }
-    public static int dividir(int a, int b) {
-        return a / b;
+    public static float dividir(int a, int b) {
+        return (float) a / b;
     }
